+++ bindings/nostr-sdk-ffi/python/examples/tags.py
@@ -10,7 +10,7 @@ tag = Tag.from_standardized(TagStandard.PUBLIC_KEY_TAG(other_user_pk, None, None
 # OR
 tag = Tag.public_key(other_user_pk)
 
-event = EventBuilder.text_note("New note from Rust Nostr python bindings", [tag]).sign_with_keys(keys)
+event = EventBuilder.text_note("New note from Rust Nostr python bindings").tags([tag]).sign_with_keys(keys)
 print(event.as_json())
 
 print("\nTags:")