+++ crates/nostr-database/examples/helper.rs
@@ -27,17 +27,15 @@ async fn main() {
     let helper = DatabaseHelper::unbounded();
 
     for i in 0..100_000 {
-        let event = EventBuilder::text_note(format!("Event #{i}"), [])
+        let event = EventBuilder::text_note(format!("Event #{i}"))
             .sign_with_keys(&keys_a)
             .unwrap();
         helper.index_event(&event).await;
 
-        let event = EventBuilder::text_note(
-            format!("Reply to event #{i}"),
-            [Tag::event(event.id), Tag::public_key(event.pubkey)],
-        )
-        .sign_with_keys(&keys_b)
-        .unwrap();
+        let event = EventBuilder::text_note(format!("Reply to event #{i}"))
+            .tags([Tag::event(event.id), Tag::public_key(event.pubkey)])
+            .sign_with_keys(&keys_b)
+            .unwrap();
         helper.index_event(&event).await;
     }
 
@@ -50,13 +48,10 @@ async fn main() {
     }
 
     for i in 0..500_000 {
-        let event = EventBuilder::new(
-            Kind::Custom(123),
-            "Custom with d tag",
-            [Tag::identifier(format!("myid{i}"))],
-        )
-        .sign_with_keys(&keys_a)
-        .unwrap();
+        let event = EventBuilder::new(Kind::Custom(123), "Custom with d tag")
+            .tag(Tag::identifier(format!("myid{i}")))
+            .sign_with_keys(&keys_a)
+            .unwrap();
         helper.index_event(&event).await;
     }