+++ crates/nostr-sdk/examples/client.rs
@@ -18,18 +18,18 @@ async fn main() -> Result<()> {
     client.connect().await;
 
     // Publish a text note
-    let builder = EventBuilder::text_note("Hello world", []);
+    let builder = EventBuilder::text_note("Hello world");
     let output = client.send_event_builder(builder).await?;
     println!("Event ID: {}", output.id().to_bech32()?);
     println!("Sent to: {:?}", output.success);
     println!("Not sent to: {:?}", output.failed);
 
     // Create a text note POW event to relays
-    let builder = EventBuilder::text_note("POW text note from rust-nostr", []).pow(20);
+    let builder = EventBuilder::text_note("POW text note from rust-nostr").pow(20);
     client.send_event_builder(builder).await?;
 
     // Send a text note POW event to specific relays
-    let builder = EventBuilder::text_note("POW text note from rust-nostr 16", []).pow(16);
+    let builder = EventBuilder::text_note("POW text note from rust-nostr 16").pow(16);
     client
         .send_event_builder_to(["wss://relay.damus.io", "wss://relay.rip"], builder)
         .await?;