+++ crates/nostr/src/event/mod.rs
@@ -471,7 +471,7 @@ mod tests {
     #[cfg(feature = "std")]
     fn test_custom_kind() {
         let keys = Keys::generate();
-        let e: Event = EventBuilder::new(Kind::Custom(123), "my content", [])
+        let e: Event = EventBuilder::new(Kind::Custom(123), "my content")
             .sign_with_keys(&keys)
             .unwrap();
 
@@ -487,10 +487,10 @@ mod tests {
     #[cfg(feature = "std")]
     fn test_event_expired() {
         let my_keys = Keys::generate();
-        let event =
-            EventBuilder::text_note("my content", [Tag::expiration(Timestamp::from(1600000000))])
-                .sign_with_keys(&my_keys)
-                .unwrap();
+        let event = EventBuilder::text_note("my content")
+            .tags([Tag::expiration(Timestamp::from(1600000000))])
+            .sign_with_keys(&my_keys)
+            .unwrap();
 
         assert!(&event.is_expired());
     }
@@ -502,12 +502,10 @@ mod tests {
         let expiry_date: u64 = now.as_u64() * 2;
 
         let my_keys = Keys::generate();
-        let event = EventBuilder::text_note(
-            "my content",
-            [Tag::expiration(Timestamp::from(expiry_date))],
-        )
-        .sign_with_keys(&my_keys)
-        .unwrap();
+        let event = EventBuilder::text_note("my content")
+            .tags([Tag::expiration(Timestamp::from(expiry_date))])
+            .sign_with_keys(&my_keys)
+            .unwrap();
 
         assert!(!&event.is_expired());
     }
@@ -516,7 +514,7 @@ mod tests {
     #[cfg(feature = "std")]
     fn test_event_without_expiration_tag() {
         let my_keys = Keys::generate();
-        let event = EventBuilder::text_note("my content", [])
+        let event = EventBuilder::text_note("my content")
             .sign_with_keys(&my_keys)
             .unwrap();
         assert!(!&event.is_expired());