+++ crates/nostr/src/nips/nip47.rs
@@ -521,12 +521,9 @@ impl Request {
     pub fn to_event(self, uri: &NostrWalletConnectURI) -> Result<Event, Error> {
         let encrypted = nip04::encrypt(&uri.secret, &uri.public_key, self.as_json())?;
         let keys: Keys = Keys::new(uri.secret.clone());
-        Ok(EventBuilder::new(
-            Kind::WalletConnectRequest,
-            encrypted,
-            [Tag::public_key(uri.public_key)],
-        )
-        .sign_with_keys(&keys)?)
+        Ok(EventBuilder::new(Kind::WalletConnectRequest, encrypted)
+            .tag(Tag::public_key(uri.public_key))
+            .sign_with_keys(&keys)?)
     }
 }