+++ crates/nostr/src/nips/nip57.rs
@@ -273,7 +273,9 @@ pub fn anonymous_zap_request(data: ZapRequestData) -> Result<Event, Error> {
     tags.push(Tag::from_standardized_without_cell(TagStandard::Anon {
         msg: None,
     }));
-    Ok(EventBuilder::new(Kind::ZapRequest, message, tags).sign_with_keys(&keys)?)
+    Ok(EventBuilder::new(Kind::ZapRequest, message)
+        .tags(tags)
+        .sign_with_keys(&keys)?)
 }
 
 /// Create **private** zap request
@@ -307,7 +309,8 @@ where
     if let Some(event_id) = data.event_id {
         tags.push(Tag::event(event_id));
     }
-    let msg: String = EventBuilder::new(Kind::ZapPrivateMessage, &data.message, tags)
+    let msg: String = EventBuilder::new(Kind::ZapPrivateMessage, &data.message)
+        .tags(tags)
         .sign_with_ctx(secp, rng, supplier, keys)?
         .as_json();
     let msg: String = encrypt_private_zap_message(rng, &secret_key, &data.public_key, msg)?;
@@ -318,7 +321,8 @@ where
         msg: Some(msg),
     }));
     let private_zap_keys: Keys = Keys::new_with_ctx(secp, secret_key);
-    Ok(EventBuilder::new(Kind::ZapRequest, "", tags)
+    Ok(EventBuilder::new(Kind::ZapRequest, "")
+        .tags(tags)
         .custom_created_at(created_at)
         .sign_with_ctx(secp, rng, supplier, &private_zap_keys)?)
 }